+++ blazor-application/Data/DataPointDistance.cs
@@ -1,0 +1,14 @@
+﻿namespace BlazorBLE.Data;
+
+public sealed class DataPointDistance
+{
+    public DataPoint DataPoint { get; }
+        
+    public double Distance { get; }
+        
+    public DataPointDistance(DataPoint dataPoint, double distance)
+    {
+        DataPoint = dataPoint;
+        Distance = distance;
+    }
+}
